enhance AboutClass
