+++ frontend/src/components/about-class-components/AboutClass.js
@@ -1,14 +1,16 @@
-import React, { useContext, useEffect, useState } from "react";
+import React, { act, useContext, useEffect, useState } from "react";
 import { glassImg } from "@/data/imagesData";
 import { CheckContext } from "@contexts/form-contexts/checkContext";
 import AboutClassInfo from "./AboutClassInfo";
 import AboutClassMenu from "./AboutClassMenu";
 import MagnifierTemporaryLabel from "../../pages/game/game-components/labels/MagnifierTemporaryLabel";
 
-const AboutClass = (props) => {
+const AboutClass = ({ level, styles, specificClass }) => {
   const [option, setOption] = useState(false);
   const { active, setActive, aboutClass, setAboutClass } =
     useContext(CheckContext);
+
+  const style = aboutClass === "container" ? styles.container : styles.bridge;
 
   const handleMenu = () => {
     setActive(false);
@@ -19,40 +21,46 @@ const AboutClass = (props) => {
     setActive(false);
   }, []);
 
+  console.log({
+    "about level": level,
+    styles: styles,
+    "what is this? ": aboutClass,
+    active,
+  });
+
   return (
-    <>
+    <section className="about-class-container">
       <img
-        className={
-          props?.specificClass ? `glass ${props.specificClass}` : "glass"
-        }
+        className={specificClass ? `glass ${specificClass}` : "glass"}
         src={glassImg}
         alt=""
         onClick={() => handleMenu()}
       />
-      <div
-        className={
-          props?.specificClass
-            ? `about-class ${props.specificClass}`
-            : "about-class"
-        }
-      >
-        {option && (
+      {option && (
+        <div
+          className={
+            specificClass
+              ? `about-class ${specificClass} ${option && "about-open"}`
+              : `about-class ${option && "about-open"}`
+          }
+        >
           <AboutClassMenu
             active={active}
             setActive={setActive}
             setAboutClass={setAboutClass}
           />
-        )}
-        {active && (
-          <AboutClassInfo
-            active={active}
-            name={`.${aboutClass}`}
-            information={props?.style}
-          />
-        )}
-      </div>
-      {props?.level === 1 ? <MagnifierTemporaryLabel option={option} /> : null}
-    </>
+
+          {active && (
+            <AboutClassInfo
+              active={active}
+              name={`.${aboutClass}`}
+              info={style}
+            />
+          )}
+        </div>
+      )}
+      {level === 1 ? <MagnifierTemporaryLabel option={option} /> : null}
+    </section>
   );
 };
