+++ frontend/src/components/aside/AsideBox.js
@@ -4,7 +4,7 @@ import "@styles/aside-styles.css";
 const AsideBox = ({ background, cssClass = "", children }) => {
   return (
     <aside
-      className={cssClass ? `AsideBox ${cssClass}` : "AsideBox"}
+      className={cssClass ? `aside-box ${cssClass}` : "aside-box"}
       style={{ backgroundImage: `url(${background})` }}
     >
       {children ? children : ""}
