+++ frontend/src/data/LevelData.js
@@ -402,6 +402,7 @@ export const levelData = [
       mainStyle: {},
       childrenStyle: {},
     },
+    deniedList: ["display"],
     partOfBridge: 3,
     master: true,
     correctPosition: {
