+++ frontend/src/pages/academy/data/exercises-settings.js
