+++ frontend/src/pages/profile/LevelDashboard.module.css
@@ -232,7 +232,7 @@
   display: flex;
   justify-content: center;
   align-items: center;
-  z-index: 50;
+  z-index: 20;
 }
 
 .modal {
