+++ frontend/src/pages/profile/LevelDashboard.module.scss
@@ -236,7 +236,7 @@ $text-color: rgb(50, 50, 50);
   display: flex;
   justify-content: center;
   align-items: center;
-  z-index: 50;
+  z-index: 20;
 }
 
 .modal {
