+++ frontend/src/pages/profile/components/modals/modal-styles.css
@@ -42,7 +42,7 @@
   display: grid;
   place-items: center;
   padding: var(--space-4);
-  z-index: 50;
+  z-index: 20;
   animation: overlay-show 150ms var(--ease-out);
 }
