+++ frontend/src/pages/registration/registration-style.css
@@ -249,7 +249,7 @@ button.link-button:hover {
   display: flex;
   justify-content: center;
   align-items: center;
-  z-index: 1000;
+  z-index: 10;
   backdrop-filter: blur(4px);
 }
